# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Service Booking API",
        "description": "Production-Quality Service Booking RESTful API in Laravel 12. This API serves a multi-user platform for managing service provider availability, enabling real-time booking by customers, and providing administrators with powerful reporting tools.",
        "contact": {
            "email": "admin@booking.api"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        },
        {
            "url": "/",
            "description": "Current Domain"
        }
    ],
    "paths": {
...
        "/api/v1/provider/availability": {
            "get": {
                "tags": [
                    "Availability"
                ],
                "summary": "Get provider's availability schedule",
                "description": "Get the recurring weekly availability schedule for the authenticated provider",
                "operationId": "07613217f8419d01b8e23eaaad1bf31a",
                "responses": {
                    "200": {
                        "description": "Availability schedule retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "01hw123456789"
                                                    },
                                                    "provider_id": {
                                                        "type": "string",
                                                        "example": "01hw555666777"
                                                    },
                                                    "day_of_week": {
                                                        "description": "1=Monday, 7=Sunday",
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "start_time": {
                                                        "type": "string",
                                                        "example": "09:00"
                                                    },
                                                    "end_time": {
                                                        "type": "string",
                                                        "example": "17:00"
                                                    },
                                                    "timezone": {
                                                        "type": "string",
                                                        "example": "UTC"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - Not a provider"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Availability"
                ],
                "summary": "Set provider's availability schedule",
                "description": "Create or update the recurring weekly availability schedule",
                "operationId": "9f3537762ca725f542e1da028951d17b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "entries"
                                ],
                                "properties": {
                                    "entries": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "day_of_week": {
                                                    "description": "1=Monday, 7=Sunday",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "start_time": {
                                                    "description": "HH:MM format",
                                                    "type": "string",
                                                    "example": "09:00"
                                                },
                                                "end_time": {
                                                    "description": "HH:MM format",
                                                    "type": "string",
                                                    "example": "17:00"
                                                },
                                                "timezone": {
                                                    "type": "string",
                                                    "example": "UTC"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Availability schedule saved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Saved"
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden - Not a provider"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format: Bearer {token}",
                "scheme": "bearer"
            }
        }
    },
...
}